полностью готовая таблица друзей
