+++ GeekBrains_Cource_Two/Controllers/CollectionController.swift
@@ -1,0 +1,29 @@
+//
+//  CollectionController.swift
+//  GeekBrains_Cource_Two
+//
+//  Created by Samborsky on 25.05.2022.
+//
+
+import UIKit
+
+class CollectionController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
